GEN-3620: fix platform fee to float and update redirect finance menu

+++ src/components/features/events/detail/info/index.tsx
@@ -110,7 +110,7 @@ const EventField = ({
       case 'feeThresholds':
         const platformFee = updateRequest.feeThresholds?.[0]?.platformFee;
         if (!platformFee) return '-';
-        const fee = parseInt(platformFee);
+        const fee = parseFloat(platformFee);
         return fee < 100 ? `${fee}%` : `Rp ${fee.toLocaleString()}`;
       case 'login_required':
         return updateRequest[fieldKey] ? 'Yes' : 'No';
@@ -524,7 +524,7 @@ export const EventDetailInfo = ({ eventDetail, showRejectionInfo = false, readOn
                   value={(() => {
                     const platformFeeValue = eventDetail.feeThresholds?.[0]?.platformFee;
                     if (!platformFeeValue) return '-';
-                    const fee = parseInt(platformFeeValue);
+                    const fee = parseFloat(platformFeeValue);
                     return fee < 100 ? `${fee}%` : `Rp ${fee.toLocaleString()}`;
                   })()}
                   isRejected={false}
